Refactor: avoid direct mutation of opinion variable
he code has been refactoured to avoid direct modification of the opinion variable. Instead, a new variable displayOpinion has been introduced, which stores a shortened version of the opinion if the allowed length is exceeded. This change improves the readability of the code and prevents potential problems resulting from data mutation.